requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
crashtest==0.4.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
distlib==0.3.9
dulwich==0.21.7
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
httptools==0.6.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
installer==0.7.0
ipykernel==6.29.5
ipython==8.32.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==24.3.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.11.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.2
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tifffile==2025.2.18
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.2.18.16
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.99
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
zipp==3.21.0
